Abstracted out common Store structures from ArtifactStore, so it is easier to reutilize them later for things like the catalog store.
