+++ mlte/store/artifact/underlying/http.py
@@ -23,8 +23,12 @@ from mlte.context.model import (
     Version,
     VersionCreate,
 )
-from mlte.store.base import Store, StoreSession, StoreURI
-from mlte.store.query import Query
+from mlte.store.artifact.artifact_store import (
+    ArtifactStore,
+    ArtifactStoreSession,
+)
+from mlte.store.artifact.query import Query
+from mlte.store.base import StoreURI
 from mlte.web.store.api.model import WriteArtifactRequest
 
 # -----------------------------------------------------------------------------
@@ -81,7 +85,7 @@ class RequestsClient(RemoteHttpStoreClient):
 # -----------------------------------------------------------------------------
 
 
-class RemoteHttpStoreSession(StoreSession):
+class RemoteHttpStoreSession(ArtifactStoreSession):
     """An in-memory implementation of the MLTE artifact store."""
 
     def __init__(self, *, url: str, client: RemoteHttpStoreClient) -> None:
@@ -302,7 +306,7 @@ def raise_for_response(response: requests.Response) -> None:
         raise errors.InternalError(f"{response.json()}")
 
 
-class RemoteHttpStore(Store):
+class RemoteHttpStore(ArtifactStore):
     """A remote HTTP implementation of the MLTE artifact store."""
 
     def __init__(
