+++ mlte/store/base.py
@@ -1,24 +1,12 @@
 """
-mlte/store/backend/base.py
+mlte/store/base.py
 
-MLTE artifact store interface implementation.
+MLTE general store interface.
 """
 
 from __future__ import annotations
 
 from enum import Enum
-from typing import List
-
-from mlte.artifact.model import ArtifactModel
-from mlte.context.model import (
-    Model,
-    ModelCreate,
-    Namespace,
-    NamespaceCreate,
-    Version,
-    VersionCreate,
-)
-from mlte.store.query import Query
 
 # -----------------------------------------------------------------------------
 # StoreType
@@ -26,7 +14,7 @@ from mlte.store.query import Query
 
 
 class StoreType(Enum):
-    """Represents the type of the MLTE artifact store."""
+    """Represents the type of an MLTE store."""
 
     LOCAL_MEMORY = "local_memory"
     """An in-memory store implementation."""
@@ -44,7 +32,7 @@ class StoreType(Enum):
 
 
 class StoreURI:
-    """Represents the URI for an artifact store instance."""
+    """Represents the URI for an store instance."""
 
     def __init__(self, uri: str, type: StoreType):
         """
@@ -92,251 +80,6 @@ class StoreSession:
             "Cannot invoke method on abstract StoreSession."
         )
 
-    # -------------------------------------------------------------------------
-    # Interface: Context
-    # -------------------------------------------------------------------------
-
-    def create_namespace(self, namespace: NamespaceCreate) -> Namespace:
-        """
-        Create a MLTE namespace.
-        :param namespace: The namespace create model
-        :return: The created namespace
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def read_namespace(self, namespace_id: str) -> Namespace:
-        """
-        Read a MLTE namespace.
-        :param namespace_id: The identifier for the namespace
-        :return: The namespace model
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def list_namespaces(self) -> List[str]:
-        """
-        List all MLTE namespaces.
-        :return: A collection of identifiers for all MLTE namespaces
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def delete_namespace(self, namespace_id: str) -> Namespace:
-        """
-        Delete a MLTE namespace.
-        :param namespace_id: The identifier for the namespace
-        :return: The deleted namespace
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def create_model(self, namespace_id: str, model: ModelCreate) -> Model:
-        """
-        Create a MLTE model.
-        :param namespace_id: The identifier for the namespace
-        :param model: The model create model
-        :return: The created model
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def read_model(self, namespace_id: str, model_id: str) -> Model:
-        """
-        Read a MLTE model.
-        :param namespace_id: The identifier for the namespace
-        :param model_id: The identifier for the model
-        :return: The model model
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def list_models(self, namespace_id: str) -> List[str]:
-        """
-        List all MLTE models in the given namespace.
-        :param namespace_id: The identifier for the namespace
-        :return: A collection of identifiers for all MLTE models
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def delete_model(self, namespace_id: str, model_id: str) -> Model:
-        """
-        Delete a MLTE model.
-        :param namespace_id: The identifier for the namespace
-        :param model_id: The identifier for the model
-        :return: The deleted model
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def create_version(
-        self, namespace_id: str, model_id: str, version: VersionCreate
-    ) -> Version:
-        """
-        Create a MLTE model version.
-        :param namespace_id: The identifier for the namespace
-        :param model_id: The identifier for the model
-        :param version: The version create model
-        :return: The created version
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def list_versions(self, namespace_id: str, model_id: str) -> List[str]:
-        """
-        List all MLTE versions in the given namespace and model.
-        :param namespace_id: the identifier for the namespace
-        :param model_id: The identifier for the model
-        :return: A collection of identifiers for all MLTE versions
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def read_version(
-        self, namespace_id: str, model_id: str, version_id: str
-    ) -> Version:
-        """
-        Read a MLTE model version.
-        :param namespace_id: The identifier for the namespace
-        :param model_id: The identifier for the model
-        :param version_id: The identifier for the model version
-        :return: The version model
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def delete_version(
-        self, namespace_id: str, model_id: str, version_id: str
-    ) -> Version:
-        """
-        Delete a MLTE model version.
-        :param namespace_id: The identifier for the namespace
-        :param model_id: The identifier for the model
-        :param version_id: The identifier for the model version
-        :return: The deleted version
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    # -------------------------------------------------------------------------
-    # Interface: Artifact
-    # -------------------------------------------------------------------------
-
-    def write_artifact(
-        self,
-        namespace_id: str,
-        model_id: str,
-        version_id: str,
-        artifact: ArtifactModel,
-        *,
-        force: bool = False,
-        parents: bool = False,
-    ) -> ArtifactModel:
-        """
-        Write an artifact.
-        :param namespace_id: The identifier for the namespace
-        :param model_id: The identifier for the model
-        :param version_id: The identifier for the model version
-        :param artifact: The artifact
-        :param force: Overwrite an artifact if it already exists
-        :param parents: Indicates whether organizational elements
-        for artifact should be implictly created (default: False)
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def read_artifact(
-        self,
-        namespace_id: str,
-        model_id: str,
-        version_id: str,
-        artifact_id: str,
-    ) -> ArtifactModel:
-        """
-        Read an artifact.
-        :param namespace_id: The identifier for the namespace
-        :param model_id: The identifier for the model
-        :param version_id: The identifier for the model version
-        :param artifact_id: The artifact identifier
-        :return: The artifact
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def read_artifacts(
-        self,
-        namespace_id: str,
-        model_id: str,
-        version_id: str,
-        limit: int = 100,
-        offset: int = 0,
-    ) -> List[ArtifactModel]:
-        """
-        Read artifacts withi limit and offset.
-        :param namespace_id: The identifier for the namespace
-        :param model_id: The identifier for the model
-        :param version_id: The identifier for the model version
-        :param limit: The limit on artifacts to read
-        :param offset: The offset on artifacts to read
-        :return: The read artifacts
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def search_artifacts(
-        self,
-        namespace_id: str,
-        model_id: str,
-        version_id: str,
-        query: Query = Query(),
-    ) -> List[ArtifactModel]:
-        """
-        Read a collection of artifacts, optionally filtered.
-        :param namespace_id: The identifier for the namespace
-        :param model_id: The identifier for the model
-        :param version_id: The identifier for the model version
-        :param query: The artifact query to apply
-        :return: A collection of artifacts that satisfy the filter
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
-    def delete_artifact(
-        self,
-        namespace_id: str,
-        model_id: str,
-        version_id: str,
-        artifact_id: str,
-    ) -> ArtifactModel:
-        """
-        Delete an artifact.
-        :param namespace_id: The identifier for the namespace
-        :param model_id: The identifier for the model
-        :param version_id: The identifier for the model version
-        :param artifact_id: The artifact identifier
-        :return: The deleted artifact
-        """
-        raise NotImplementedError(
-            "Cannot invoke method on abstract StoreSession."
-        )
-
 
 # -----------------------------------------------------------------------------
 # Store
@@ -345,7 +88,7 @@ class StoreSession:
 
 class Store:
     """
-    An abstract
+    An abstract store.
 
     A Store instance is the "static" part of a store configuration.
     In contrast, a StoreSession represents an active session with the store.
@@ -358,7 +101,7 @@ class Store:
         """
 
         self.uri = uri
-        """The parsed artifact store URI."""
+        """The parsed store URI."""
 
     def session(self) -> StoreSession:
         """
