+++ mlte/web/store/api/endpoints/artifact.py
@@ -13,7 +13,7 @@ from fastapi import APIRouter, HTTPException
 import mlte.store.error as errors
 import mlte.web.store.api.codes as codes
 from mlte.artifact.model import ArtifactModel
-from mlte.store.query import Query
+from mlte.store.artifact.query import Query
 from mlte.web.store.api import dependencies
 from mlte.web.store.api.model import WriteArtifactRequest, WriteArtifactResponse
