+++ mlte/web/store/main.py
@@ -14,8 +14,8 @@ from pydantic.networks import HttpUrl
 
 import mlte.web.store.app_factory as app_factory
 import mlte.web.store.util.origins as util
+from mlte.store.artifact.factory import create_store
 from mlte.store.base import StoreType
-from mlte.store.factory import create_store
 from mlte.web.store.api.api import api_router
 from mlte.web.store.core.config import settings
 from mlte.web.store.state import state
