+++ test/fixture/store.py
@@ -12,8 +12,11 @@ import pytest
 
 from mlte.context.context import Context
 from mlte.context.model import ModelCreate, NamespaceCreate, VersionCreate
-from mlte.store.base import ManagedSession, Store
-from mlte.store.factory import create_store
+from mlte.store.artifact.artifact_store import (
+    ArtifactStore,
+    ManagedArtifactSession,
+)
+from mlte.store.artifact.factory import create_store
 
 # The namespace identifier for default context
 FX_NAMESPACE_ID = "ns0"
@@ -26,16 +29,16 @@ FX_VERSION_ID = "v0"
 
 
 @pytest.fixture(scope="function")
-def store() -> Store:
+def store() -> ArtifactStore:
     """Create an in-memory artifact store."""
     return create_store("memory://")
 
 
 @pytest.fixture(scope="function")
-def store_with_context() -> Tuple[Store, Context]:
+def store_with_context() -> Tuple[ArtifactStore, Context]:
     """Create an in-memory artifact store with initial context."""
     store = create_store("memory://")
-    with ManagedSession(store.session()) as handle:
+    with ManagedArtifactSession(store.session()) as handle:
         _ = handle.create_namespace(NamespaceCreate(identifier=FX_NAMESPACE_ID))
         _ = handle.create_model(
             FX_NAMESPACE_ID, ModelCreate(identifier=FX_MODEL_ID)
