+++ test/measurement/cpu/test_local_process_cpu_utilization.py
@@ -18,7 +18,7 @@ from mlte.context.context import Context
 from mlte.evidence.metadata import EvidenceMetadata, Identifier
 from mlte.measurement.cpu import CPUStatistics, LocalProcessCPUUtilization
 from mlte.spec.condition import Condition
-from mlte.store.base import Store
+from mlte.store.artifact.artifact_store import ArtifactStore
 from mlte.validation.result import Failure, Success
 
 from ...fixture.store import store_with_context  # noqa
@@ -118,7 +118,7 @@ def test_cpu_windows_evaluate() -> None:
     is_windows(), reason="LocalProcessCPUUtilization not supported on Windows."
 )
 def test_result_save_load(
-    store_with_context: Tuple[Store, Context]  # noqa
+    store_with_context: Tuple[ArtifactStore, Context]  # noqa
 ) -> None:
     store, ctx = store_with_context
