+++ test/web/store/fixure/http.py
@@ -13,7 +13,7 @@ from fastapi.testclient import TestClient
 
 import mlte.web.store.app_factory as app_factory
 from mlte.artifact.type import ArtifactType
-from mlte.store.factory import create_store
+from mlte.store.artifact.factory import create_store
 from mlte.web.store.api.api import api_router
 from mlte.web.store.core.config import settings
 from mlte.web.store.state import state
